refactor: created pipeline folder
