+++ packages/developer-portal/src/components/pages/apps/page/index.tsx
@@ -26,7 +26,7 @@ import { useAppState } from '../state/use-app-state'
 import { getCurrentPage } from '../utils/get-current-page'
 import { Helper } from './helper'
 import { cx } from '@linaria/core'
-import { AppPipeline } from '../edit/app-pipeline'
+import { AppPipeline } from '../pipeline/app-pipeline'
 import { selectLoginIdentity } from '@/selector/auth'
 import { useReapitConnect } from '@reapit/connect-session'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
